--- app/views/spaces/edit/_url.html.erb
@@ -1,1 +1,0 @@
-<%= form.input :url %>
