+++ app/views/spaces/new.html.erb
@@ -8,7 +8,6 @@
       <div data-duplicate-checker-target="duplicatesRenderHere"></div>
       <div>
           <%= edit_field 'space_type', form %>
-          <%= edit_field 'url', form %>
           <%= edit_field 'space_group', form %>
           <span class="flex gap-4 items-baseline mt-4">
           <%= form.button :submit %>
