+++ spec/views/reviews/index.html.erb_spec.rb
@@ -12,7 +12,6 @@ RSpec.describe "reviews/index.html.erb", type: :view do
     expect(rendered).to match(/h1/)
     expect(rendered).to match(/#{reviews.first.comment}/)
     expect(rendered).to match(/#{reviews.second.comment}/)
-    expect(rendered).to match(/#{reviews.second.price}/)
     expect(rendered).to match(/#{reviews.third.star_rating}/)
   end
 end
